[NCL-2662] Extract functionality from Reports REST endpoint to ReportsFacade
